% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_sample}
\alias{boot_sample}
\title{Create bootstrap samples form INAR parameters.}
\usage{
boot_sample(Y, boot_param)
}
\arguments{
\item{Y}{A n x d matrix containing token counts for n different tokens and d
different documents.}

\item{boot_param}{.}
}
\value{
.
}
\description{
Create bootstrap samples form INAR parameters.
}
