% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.R
\name{estim}
\alias{estim}
\title{Estimate Poisson parameter matrix with reduced rank.}
\usage{
estim(Y, K, tol = 1e-05, verbose = TRUE)
}
\arguments{
\item{Y}{A n x d matrix containing words counts for n different words and d
different documents.}

\item{K}{A positive integer. The rank of the normalized parameter matrix.}
}
\value{
A n x d matrix. Contains the logs of the Poisson parameters.
}
\description{
Estimate Poisson parameter matrix with reduced rank.
}
