% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{estim_INAR}
\alias{estim_INAR}
\title{Estimate INAR parameters.}
\usage{
estim_INAR(
  Y,
  theta,
  group = rep("default", nrow(Y)),
  time = seq_len(group),
  row_ids = rownames(Y)
)
}
\arguments{
\item{Y}{A n x d matrix containing token counts for n different tokens and d
different documents.}

\item{theta}{A n x d matrix. The logs of the Poisson parameters.}

\item{group}{A character vector of length n for grouping.}

\item{time}{A numeric vector of length n to order elements by time.}

\item{row_ids}{A index vector of length n mapping the ordering of group and
time to the ordering in the rows of theta.}
}
\value{
A named list with INAR parameters for each group.
}
\description{
Estimate INAR parameters.
}
