% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{estim_INAR_token}
\alias{estim_INAR_token}
\title{Estimate the parameters of the INAR model for one token}
\usage{
estim_INAR_token(y, theta)
}
\arguments{
\item{y}{counts of word for each point in time}

\item{theta}{parameter vector for word for each point in time}
}
\description{
Estimate the parameters of the INAR model for one token
}
