% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.R
\name{plist2theta}
\alias{plist2theta}
\title{Combine a list of parameter vectors to the parameter matrix.}
\usage{
plist2theta(plist)
}
\arguments{
\item{plist}{A list of parameter vectors as returned by theta2plist().}
}
\value{
A n x d matrix. Contains the logs of the Poisson parameters.
}
\description{
Combine a list of parameter vectors to the parameter matrix.
}
