% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisrrr-package.R
\docType{package}
\name{poisrrr-package}
\alias{poisrrr}
\alias{poisrrr-package}
\title{poisrrr: Poisson Reduced-Rank Models}
\description{
Implements algorithms to fit the Poisson reduced-rank models of C. 
    Jentsch, E. Mammen and E. R. Lee (2020) Poisson reduced rank models with an 
    application to political text data. Biometrika.
}
\author{
\strong{Maintainer}: Christof Schötz \email{schoetz@math.uni-heidelberg.de}

Authors:
\itemize{
  \item Eun Ryung Lee \email{erlee@skku.edu}
  \item Carsten  Jentsch \email{jentsch@statistik.tu-dortmund.de}
  \item Enno Mammen \email{mammen@math.uni-heidelberg.de}
}

}
\keyword{The}
\keyword{internal}
\keyword{package.}
\keyword{poisrrr}
