% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate_fit}
\alias{rotate_fit}
\title{Rotate a point configuration to fit another one.}
\usage{
rotate_fit(x, target)
}
\arguments{
\item{x}{A numeric n x d matrix.}

\item{target}{A numeric n x d matrix.}
}
\value{
A n x d matrix containing the rotated points of x which minimize the
squared Euclidean distance to target.
}
\description{
Rotate a point configuration to fit another one.
}
