% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.R
\name{theta2plist}
\alias{theta2plist}
\title{Extract parameter vectors from parameter matrix.}
\usage{
theta2plist(theta, K)
}
\arguments{
\item{theta}{A n x d matrix. The logs of the Poisson parameteres.}

\item{K}{A positive integer. The rank of the normalized parameter matrix.}
}
\value{
A list containing the normalizing values and the SVD of the
normalized matrix.
}
\description{
Extract parameter vectors from parameter matrix.
}
