% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood}
\usage{
loglik(Y, theta)
}
\arguments{
\item{Y}{A n x d matrix containing words counts for n different words and d
different documents.}

\item{theta}{A n x d matrix. The logs of the Poisson parameteres.}
}
\value{
A number. The Log-likelihood.
}
\description{
Log-likelihood
}
